Implement inital folder restructure
