+++ auth/admin_hash.php
