+++ frontend/dashboard.html
@@ -21,7 +21,7 @@
         <nav class="navigation">
             <ul class="list-items">
                 <li><img src="/docs/Home.svg" alt=""><a href="">Home</a></li>
-                <li><img src="/docs/new-svgrepo-com.svg" alt=""><a href="">New Contact</a></li>
+                <li><img src="/assets/new-svgrepo-com.svg" alt=""><a href="">New Contact</a></li>
                 <li><img src="/docs/MySpace.svg" alt=""><a href="">Users</a></li>
                 <hr>
                 <li><img src="/docs/logout-svgrepo-com.svg" alt=""><a href="../backend/logout.php">Logout</a></li>
